@@ -101,105 +101,122 @@
     }
   },
   "devDependencies": {
-    "asar": "^0.14.3",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.5",
-    "babel-jest": "^23.2.0",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-add-module-exports": "^0.2.1",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-do-expressions": "^7.0.0",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-bind": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/plugin-transform-classes": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "asar": "^1.0.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^23.4.2",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-add-module-exports": "^1.0.0",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-dynamic-import-webpack": "^1.0.2",
-    "babel-plugin-flow-runtime": "^0.17.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-es2015-classes": "^6.24.1",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
+    "babel-plugin-dynamic-import-webpack": "^1.1.0",
+    "babel-plugin-flow-runtime": "^0.18.0",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-react-optimize": "^1.0.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.26.0",
     "babili-webpack-plugin": "^0.1.2",
     "boiler-room-custodian": "^0.6.2",
-    "chalk": "^2.4.1",
-    "concurrently": "^3.6.0",
+    "chalk": "^2.4.2",
+    "concurrently": "^4.1.0",
     "cross-env": "^5.2.0",
     "cross-spawn": "^6.0.5",
-    "css-loader": "^0.28.11",
+    "css-loader": "^2.1.1",
     "devtron": "^1.4.0",
-    "electron": "^2.0.4",
-    "electron-builder": "^20.19.2",
+    "electron": "^4.1.0",
+    "electron-builder": "^20.39.0",
     "electron-devtools-installer": "^2.2.4",
-    "enzyme": "^3.3.0",
-    "eslint": "^5.0.1",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-formatter-pretty": "^1.3.0",
-    "eslint-import-resolver-webpack": "^0.10.1",
-    "eslint-plugin-compat": "^2.4.0",
-    "eslint-plugin-flowtype": "^2.49.3",
-    "eslint-plugin-flowtype-errors": "^3.6.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-jest": "^21.17.0",
-    "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-react": "^7.10.0",
-    "express": "^4.16.3",
+    "enzyme": "^3.9.0",
+    "eslint": "^5.15.3",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-formatter-pretty": "^2.1.1",
+    "eslint-import-resolver-webpack": "^0.11.0",
+    "eslint-plugin-compat": "^3.0.1",
+    "eslint-plugin-flowtype": "^3.4.2",
+    "eslint-plugin-flowtype-errors": "^4.0.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jest": "^22.4.1",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-react": "^7.12.4",
+    "express": "^4.16.4",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "fbjs-scripts": "^0.8.3",
-    "file-loader": "^1.1.11",
-    "flow-bin": "^0.75.0",
+    "fbjs-scripts": "^1.1.0",
+    "file-loader": "^3.0.1",
+    "flow-bin": "^0.95.1",
     "flow-runtime": "^0.17.0",
-    "flow-typed": "^2.5.0",
+    "flow-typed": "^2.5.1",
     "html-webpack-plugin": "^3.2.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.3.0",
-    "jsdom": "^11.11.0",
-    "mini-css-extract-plugin": "^0.4.1",
+    "jest": "^24.5.0",
+    "jsdom": "^14.0.0",
+    "mini-css-extract-plugin": "^0.5.0",
     "minimist": "^1.2.0",
-    "node-sass": "^4.9.1",
+    "node-sass": "^4.11.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-test-renderer": "^16.4.1",
+    "react-test-renderer": "^16.8.4",
     "redux-logger": "^3.0.6",
-    "sass-loader": "^7.0.3",
-    "sinon": "^6.1.2",
-    "spectron": "^3.8.0",
-    "style-loader": "^0.21.0",
-    "stylelint": "^9.3.0",
+    "sass-loader": "^7.1.0",
+    "sinon": "^7.2.7",
+    "spectron": "^5.0.0",
+    "style-loader": "^0.23.1",
+    "stylelint": "^9.10.1",
     "stylelint-config-standard": "^18.2.0",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.15.1",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.3"
+    "uglifyjs-webpack-plugin": "^2.1.2",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.29.6",
+    "webpack-bundle-analyzer": "^3.1.0",
+    "webpack-cli": "^3.3.0",
+    "webpack-dev-server": "^3.2.1",
+    "webpack-merge": "^4.2.1",
+    "@babel/preset-flow": "^7.0.0"
   },
   "dependencies": {
     "@exponent/electron-cookies": "^2.0.0",
-    "apollo-cache-inmemory": "^1.2.5",
-    "apollo-client": "^2.3.5",
-    "apollo-link-http": "^1.5.4",
+    "apollo-cache-inmemory": "^1.5.1",
+    "apollo-client": "^2.5.1",
+    "apollo-link-http": "^1.5.14",
     "classnames": "^2.2.6",
-    "electron-debug": "^2.0.0",
+    "connected-react-router": "^6.3.2",
+    "electron-debug": "^2.1.0",
     "electron-store": "^2.0.0",
-    "graphql-tag": "^2.9.2",
-    "history": "^4.7.2",
-    "marked": "^0.4.0",
-    "mousetrap": "^1.6.2",
+    "feed": "^2.0.4",
+    "graphql-tag": "^2.10.1",
+    "history": "^4.9.0",
+    "marked": "^0.6.1",
+    "mousetrap": "^1.6.3",
     "persistgraphql": "^0.3.11",
-    "react": "^16.4.1",
-    "react-apollo": "^2.1.9",
-    "react-dom": "^16.4.1",
-    "react-hot-loader": "^4.3.3",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "react-timeago": "^4.1.9",
-    "redux": "^4.0.0",
-    "redux-form": "^7.4.2",
+    "react": "^16.8.4",
+    "react-apollo": "^2.5.2",
+    "react-dom": "^16.8.4",
+    "react-hot-loader": "^4.8.0",
+    "react-redux": "^6.0.1",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
+    "react-timeago": "^4.4.0",
+    "redux": "^4.0.1",
+    "redux-form": "^8.1.0",
     "redux-thunk": "^2.3.0",
-    "source-map-support": "^0.5.6"
+    "source-map-support": "^0.5.11"
   },
   "devEngines": {
     "node": ">=6.x",
