@@ -1,6 +1,11 @@
 // @flow
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.production'); // eslint-disable-line global-require
-} else {
-  module.exports = require('./configureStore.development'); // eslint-disable-line global-require
-}
+import configureStoreDev from './configureStore.development';
+import configureStoreProd from './configureStore.production';
+
+const selectedConfigureStore = process.env.NODE_ENV === 'production'
+  ? configureStoreProd
+  : configureStoreDev;
+
+export const { configureStore } = selectedConfigureStore;
+
+export const { history } = selectedConfigureStore;
