@@ -1,18 +1,18 @@
 // @flow
 import { combineReducers } from 'redux';
-import { routerReducer as router } from 'react-router-redux';
+import { connectRouter } from 'connected-react-router';
 import { reducer as form } from 'redux-form';
 
 import links from './links';
 import userInterface from './userInterface';
 import configuration from './configuration';
 
-const rootReducer = combineReducers({
-  router,
+const createRootReducer = (history) => combineReducers({
+  router: connectRouter(history),
   form,
   links,
   userInterface,
   configuration,
 });
 
-export default rootReducer;
+export default createRootReducer;
