@@ -1,9 +1,9 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import { createHashHistory as createHistory } from 'history';
-import { routerMiddleware, push } from 'react-router-redux';
+import { routerMiddleware, push } from 'connected-react-router';
 import { createLogger } from 'redux-logger';
-import rootReducer from '../reducers';
+import createRootReducer from '../reducers';
 import middlewares from '../middleware';
 
 
@@ -30,7 +30,6 @@ const configureStore = (initialState: ?counterStateType) => {
 
   // Redux DevTools Configuration
   const actionCreators = {
-
     push,
   };
   // If Redux DevTools Extension is installed use it, otherwise use Redux compose
@@ -49,7 +48,11 @@ const configureStore = (initialState: ?counterStateType) => {
   const enhancer = composeEnhancers(...enhancers);
 
   // Create Store
-  const store = createStore(rootReducer, initialState, enhancer);
+  const store = createStore(
+    createRootReducer(history),
+    initialState,
+    enhancer
+  );
 
   if (module.hot) {
     module.hot.accept('../reducers', () =>
