@@ -2,8 +2,8 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { createHashHistory as createHistory } from 'history';
-import { routerMiddleware } from 'react-router-redux';
-import rootReducer from '../reducers';
+import { routerMiddleware } from 'connected-react-router';
+import createRootReducer from '../reducers';
 import middlewares from '../middleware';
 
 
@@ -12,7 +12,11 @@ const router = routerMiddleware(history);
 const enhancer = applyMiddleware(thunk, router, ...middlewares);
 
 function configureStore(initialState) {
-  return createStore(rootReducer, initialState, enhancer);
+  return createStore(
+    createRootReducer(history),
+    initialState,
+    enhancer
+  );
 }
 
 export default { configureStore, history };
